- Reverted wxscintilla platform.h accidentially committed in 8511
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@8513 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/sdk/wxscintilla/src/scintilla/include/Platform.h
@@ -69,10 +69,8 @@
 namespace Scintilla {
 #endif
 
-#include <cstdint>
-
-typedef intptr_t XYPOSITION;
-typedef intptr_t XYACCUMULATOR;
+typedef float XYPOSITION;
+typedef double XYACCUMULATOR;
 //#define XYPOSITION int
 
 // Underlying the implementation of the platform classes are platform specific types.
